complaint_fullstack/backend/requirements.txt
flask
flask-cors
pyjwt
